+++ src/realm/sync/subscriptions.cpp
@@ -162,6 +162,11 @@ int64_t SubscriptionSet::version() const
     return m_version;
 }
 
+DB::version_type SubscriptionSet::snapshot_version() const
+{
+    return m_snapshot_version;
+}
+
 SubscriptionSet::State SubscriptionSet::state() const
 {
     return m_state;
@@ -704,6 +709,25 @@ SubscriptionStore::get_next_pending_version(int64_t last_query_version, DB::vers
     auto query_version = obj.get_primary_key().get_int();
     auto snapshot_version = obj.get<int64_t>(m_sub_set_snapshot_version);
     return PendingSubscription{query_version, static_cast<DB::version_type>(snapshot_version)};
+}
+
+std::vector<SubscriptionSet> SubscriptionStore::get_pending_subscriptions() const
+{
+    std::vector<SubscriptionSet> subscriptions_to_recover;
+    auto active_sub = get_active();
+    auto cur_query_version = active_sub.version();
+    DB::version_type db_version = 0;
+    if (active_sub.state() == SubscriptionSet::State::Complete) {
+        db_version = active_sub.snapshot_version();
+    }
+    REALM_ASSERT_EX(db_version != DB::version_type(-1), active_sub.state());
+    // get a copy of the pending subscription sets since the active version
+    while (auto next_pending = get_next_pending_version(cur_query_version, db_version)) {
+        cur_query_version = next_pending->query_version;
+        db_version = next_pending->snapshot_version;
+        subscriptions_to_recover.push_back(get_by_version(cur_query_version));
+    }
+    return subscriptions_to_recover;
 }
 
 MutableSubscriptionSet SubscriptionStore::get_mutable_by_version(int64_t version_id)