+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 * Allow flexible sync with discard local client resets. ([#5404](https://github.com/realm/realm-core/pull/5404))
+* Allow flexible sync with recovery client resets. ([#5562](https://github.com/realm/realm-core/issues/5562))
 
 ### Fixed
 * <How do the end-user experience this issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)